pyproject.toml
[tool.poetry]
name = "tts-app"
version = "0.1.0"
description = "a text-to-speech app"
authors = ["Yogesh <yoge.self@gmail.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
fastapi = {extras = ["standard"], version = "^0.112.2"}
uvicorn = "^0.30.6"
gtts = "^2.5.3"
sqlalchemy = "^2.0.32"
psycopg2-binary = "^2.9.9"
openai = "^1.42.0"
deepgram-sdk = "^3.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
